move new trigger tests to conformance suite
