+++ test/conformance/new_trigger_filters_test.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rekt
+package conformance
 
 import (
 	"testing"
